outsourcing prefetchStockItems from category actions
