+++ core/modules/catalog-next/store/category/actions.ts
@@ -14,6 +14,7 @@ import { configureProductAsync } from '@vue-storefront/core/modules/catalog/help
 import { DataResolver } from 'core/data-resolver/types/DataResolver';
 import { Category } from '../../types/Category';
 import { _prepareCategoryPathIds } from '../../helpers/categoryHelpers';
+import { prefetchStockItems } from '../../helpers/cacheProductsHelper';
 import chunk from 'lodash-es/chunk'
 
 const actions: ActionTree<CategoryState, RootState> = {
@@ -76,23 +77,8 @@ const actions: ActionTree<CategoryState, RootState> = {
       updateState: false // not update the product listing - this request is only for caching
     }, { root: true })
     if (products.filterUnavailableVariants && products.configurableChildrenStockPrefetchStatic) { // prefetch the stock items
-      const skus = []
-      let prefetchIndex = 0
-      cachedProductsResponse.items.map(i => {
-        if (products.configurableChildrenStockPrefetchStaticPrefetchCount > 0) {
-          if (prefetchIndex > products.configurableChildrenStockPrefetchStaticPrefetchCount) return
-        }
-        skus.push(i.sku) // main product sku to be checked anyway
-        if (i.type_id === 'configurable' && i.configurable_children && i.configurable_children.length > 0) {
-          for (const confChild of i.configurable_children) {
-            const cachedItem = rootState.stock.cache[confChild.id]
-            if (typeof cachedItem === 'undefined' || cachedItem === null) {
-              skus.push(confChild.sku)
-            }
-          }
-          prefetchIndex++
-        }
-      })
+      const skus = prefetchStockItems(cachedProductsResponse, rootState.stock.cache)
+
       for (const chunkItem of chunk(skus, 15)) {
         dispatch('stock/list', { skus: chunkItem }, { root: true }) // store it in the cache
       }
